study-buddy-backend/requirements.txt
fastapi
uvicorn
langchain==0.1.20
langchain-core==0.1.52
langchain-google-genai==0.0.11
langchain-community==0.0.38
google-generativeai
faiss-cpu
sentence-transformers
python-dotenv
